refactor(ui): integrate chat view shell extraction
